requirements.txt
bs4
requests
pyTelegramBotAPI
python-Levenshtein